Moved changes to dev version of Tajima notebook
